\name{open31114-package}
\alias{open31114-package}
\alias{open31114}
\docType{package}
\title{
Application de la norme NF S 31-114 (mesure du bruit éolien)
}
\description{
Le package rassemble les fonctions utiles au calcul des indicateurs de bruit, définis par la norme NF S 31-114 : Mesurage du bruit dans l'environnement avec et sans activité éolienne.\cr
    Sponsorisé par Maïa Eolis.
}
\details{
\tabular{ll}{
Package: \tab open-31114\cr
Type: \tab Package\cr
Version: \tab 0.2\cr
Date: \tab 2014-11-04\cr
License: \tab GPL-3\cr
}
recentrage(data.frame)\cr
effectue les interpolations et extrapolations aux valeurs de vitesses de vent entières
}
\author{
Thomas Jacoviac

Maintainer: <tjacoviac@priseauvent.fr>
}
\references{
\url{http://www.developpement-durable.gouv.fr/IMG/pdf/guide_eolien_finalv1.pdf}\cr
\url{http://www.rhone-alpes.developpement-durable.gouv.fr/IMG/pdf/Norme_Acoustique_NF_S_31-114_V_07_07_2011_v3_cle58e582.pdf}
}
\keyword{ package }
\seealso{
\url{http://www.maiaeolis.fr/}
}
\examples{
}
