% Generated by roxygen2 (4.0.2): do not edit by hand
\name{recentrage}
\alias{recentrage}
\title{Interpolations et extrapolations aux valeurs de vitesses de vent entières}
\usage{
recentrage(dfIN, classe = "classe", Vs_moyen = "Vs_moyen",
  L_median = "L_median")
}
\arguments{
\item{dfIN}{data.frame en entrée contient au moins 3 colonnes :
 \itemize{
   \item classe    : numeric classe de vent
   \item Vs_moyen  : numeric moyenne des vitesses de vent
   \item L_median  : numeric médiane des descripteurs du niveau sonore
   }
Les colonnes doivent être nommées.\cr
Par défaut, la fonction attend les colonnes 'classe', Vs_moyen' et 'L_median'.}

\item{classe}{character Optionnel}

\item{Vs_moyen}{character Optionnel}

\item{L_median}{character Optionnel

Les arguments classe, Vs_moyen et L_median permettent de spécifier le nom des colonnes
si nécessaire.}
}
\value{
data.frame en sortie contient 2 colonnes :
 \itemize{
  \item classe : numeric
  \item indicateur : numeric
}
}
\description{
Applique la norme NF S 31-114\cr
Les cas sont nommés de A à E d'après le projet de norme NFS 31114 v9 annexe B.\cr
La fonction n'a pas connaissance du nombre de descripteurs du niveau sonore
par classe de vitesse de vent. Le cas E (moins de 10 descripteurs) doit être déterminé en amont.\cr
Pour ce cas, la data.frame en entrée ne contiendra pas de ligne pour la classe considérée.
A titre transitoire, la valeur de Vs_moyen ou de L_median pourra être à NA.
}

