% Generated by roxygen2 (4.0.2): do not edit by hand
\name{recentrage}
\alias{recentrage}
\title{Interpolations et extrapolations aux valeurs de vitesses de vent entières}
\usage{
recentrage(dfIN, Vs_moyen = "Vs_moyen", L_median = "L_median")
}
\arguments{
\item{dfIN}{data.frame en entrée contient 3 colonnes :
 \itemize{
   \item classe    : numeric
   \item Vs_moyen  : numeric
   \item L_median  : numeric
   }
Les colonnes doivent être nommées.
La colonne classe est obligatoirement nommée ainsi.
Les arguments Vs_moyen et L_median permettent de passer le nom des colonnes spécifiées.}

\item{Vs_moyen}{character Optionnel}

\item{L_median}{character Optionnel}
}
\value{
dfOUT data.frame de sortie contient 2 colonnes :
 \itemize{
  \item classe : numeric
  \item indicateur : numeric
}
}
\description{
Applique la norme NF S 31-114
Les cas sont nommés d'après le projet de norme NFS 31114 v9 annexe B
}
\details{
Le cas E (moins de 10 descripteurs) ne peut être attribué par la fonction.\cr
Il appartient à l'utilisateur de ne pas transmettre à cette fonction
de données numériques pour une classe contenant moins de 10 descripteurs du niveau sonore.
}

